requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.7
flask-wtf==1.2.1
pymysql==1.1.1
boto3==1.34.131
python-dotenv==1.0.1
gunicorn==22.0.0
werkzeug==3.0.3
wtforms==3.1.2